feat(scoring): puntaje DECIMAL fiel a prod — Examen Nacional 1.25/correcta (total 100)
Prod usa 1.25 puntos/correcta en el Nacional (80 preg = 100 exacto); con points
INT se truncaba a 1 (quedaba /80). Se pasa el puntaje a float64/FLOAT en todo el
stack del score:
- exams: ExamQuestion.points/incorrect/blank + ExamAttempt.score/max_score
  (proto double, migration 016 ALTER COLUMN FLOAT, domain, repos, scoring).
- analytics: UpstreamAttempt + TestResult + ReporteEstudiante score -> float64,
  proto double.
- gateway: structs JSON de attempt/add-question -> float64 (hubspot sync castea
  a int, sync secundario que P1 ni hacia).
UCSP (5/0/1, total 450) sigue exacto. Falta UPDATE de los 47 Nacional a 1.25.

+++ analytics_service/internal/core/domain/dashboard.go
@@ -132,8 +132,8 @@ type TestResult struct {
 	// renderizar el result-card por intento: simulacro carga sus answers y
 	// vocacional/estilos su reporte por attempt_id.
 	AttemptID   AttemptID
-	Score       int32
-	MaxScore    int32
+	Score       float64
+	MaxScore    float64
 	SubmittedAt time.Time
 }
